fix(eval): address PR review feedback for structured JSON output
- Fix Anthropic adapter truncation order: validate response length
  before prepending JSON prefill character to avoid broken JSON
- Consolidate three divergent extract_json_payload implementations
  into shared json_utils module using robust brace-depth counting
- Remove duplicate implementations from semantic.py, consensus.py,
  and qa.py in favor of the shared utility
